Improvement of unit test with the spec_helper
