+++ spec/unit/_core_bot_spec.rb
@@ -1,39 +1,4 @@
-require 'slack-ruby-client'
-require './scripts/date'
-require './scripts/system'
-require './scripts/quote'
-require './scripts/ssh'
-require './scripts/pass'
-require './scripts/tc'
-require './scripts/2fa'
-require './scripts/random'
-require './scripts/horoscopo'
-require './scripts/lotery'
-require './scripts/security_check'
-require './scripts/wikipedia'
-require './scripts/celery'
-require './scripts/lingo'
-require './scripts/stock'
-require './scripts/qr'
-require './scripts/flight'
-require './scripts/weather'
-require './core'
-
-class BotValue
-  BOT_ICON = ':energon:'.freeze
-  BOT_NAME = 'ENERTEST'.freeze
-end
-
-# Acts like a incoming message from Slack
-module Fake
-  def self.text
-    MESSAGE
-  end
-
-  def self.channel
-    '#bots'
-  end
-end
+require_relative '../spec_helper'
 
 describe '#bot' do
   before(:each) do
@@ -66,7 +31,7 @@ describe '#bot' do
         'enerbot clima'
       ].each do |text|
         MESSAGE = text
-        Case.bot(Fake)
+        Case.bot(BotValue)
       end
     end
 
@@ -76,7 +41,7 @@ describe '#bot' do
         'enerbot tc masterca'
       ].each do |text|
         MESSAGE = text
-        Case.bot(Fake)
+        Case.bot(BotValue)
       end
     end
   end
